+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [6.0.0](https://github.com/alexdlaird/pyngrok/compare/5.2.3...6.0.0) - 2023-04-12
 ### Changed
-- Changed default installer behavior to download `ngrok` v3 when no version is specified.
+- Changed default installer behavior to download `ngrok` v3 by default.
 - Documentation updates.
 - Test updates.